replace build_cop_profiles and build_heat_source_utilisation_profiles with a single rule and compute heat_pump_coolling for COP approximation iteratively when enabled
