+++ scripts/build_heat_source_profiles/base_cop_approximator.py
